Split pool interface
